requirements.txt
arxiv==1.4.8
anthropic==0.8.1
PyPDF2==3.0.1
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
